backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.3
inflection==0.3.1
iniconfig==1.1.1
lazy-object-proxy==1.8.0
Markdown==3.4.1
packaging==21.3
phonenumbers==8.12.57
Pillow==9.3.0
pluggy==1.0.0
psycopg2==2.9.5
pyparsing==3.0.9
pytest==7.2.0
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-django==4.5.2
pytest-drf==1.1.3
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
wrapt==1.14.1
